code-sdc/requirements.txt
matplotlib
numpy
pandas
keras
scikit-learn
opencv
Pillow
eventlet
flask
python-socketio
